Create LICENSE
According to what you said as a reply in https://github.com/WebDevSimplified/wordle-clone/issues/3 this is the most simple open-source license to add. It is the MIT License. In the future, if you create your repo through GitHub, you can just click 'License', then pick 'MIT' and it'll be automatically added.

I add this, because even if you say it is ok to use, the code itself by default is closed source on GitHub, even though it is a public repo.